requirements.txt
# Phase 1.1: Database Foundation
sqlalchemy==2.0.23
python-dotenv==1.0.0
pytest==7.4.3

# Phase 1.2: OpenAI Integration (for later)
# openai==1.3.0
# pydantic==2.5.0

# Phase 1.3: REST API (for later)
# fastapi==0.104.1
# uvicorn[standard]==0.24.0
# python-jose[cryptography]==3.3.0
# passlib[bcrypt]==1.7.4

# Phase 1.4: CLI (for later)
# click==8.1.7
# requests==2.31.0
# tabulate==0.9.0

# Phase 1.5: Scheduler (for later)
# schedule==1.2.0
